# Conflicts:
#	src/main/java/shop/gaship/gashipshoppingmall/orderproduct/entity/OrderProduct.java
#	src/main/java/shop/gaship/gashipshoppingmall/orderproduct/repository/impl/OrderProductRepositoryImpl.java
#	src/main/resources/application-dev.properties
#	src/main/resources/application-prod.properties

+++ src/main/java/shop/gaship/gashipshoppingmall/totalsale/dto/response/TotalSaleResponseDto.java
@@ -19,6 +19,6 @@ public class TotalSaleResponseDto {
     private Long orderCancelCnt;
     private Long orderSaleCnt;
     private Long totalAmount;
-    private Integer cancelAmount;
+    private Long cancelAmount;
     private Long orderSaleAmount;
 }
